from QuantStack to DerThorsten master

+++ test/test_xaxis_iterator.cpp
@@ -29,7 +29,7 @@ namespace xt
     {
         xarray<int> a = get_test_array();
         auto iter_begin = axis_begin(a);
-        EXPECT_EQ(2, (*iter_begin).dimension());
+        EXPECT_EQ(size_t(2), (*iter_begin).dimension());
         EXPECT_EQ(a.shape()[1], (*iter_begin).shape()[0]);
         EXPECT_EQ(a.shape()[2], (*iter_begin).shape()[1]);
 
@@ -44,7 +44,7 @@ namespace xt
         auto iter = axis_begin(a);
         ++iter;
 
-        EXPECT_EQ(2, (*iter).dimension());
+        EXPECT_EQ(size_t(2), (*iter).dimension());
         EXPECT_EQ(a.shape()[1], (*iter).shape()[0]);
         EXPECT_EQ(a.shape()[2], (*iter).shape()[1]);
 
@@ -69,7 +69,7 @@ namespace xt
         ++iter;
         auto niter = axis_begin(*iter);
         ++niter;
-        EXPECT_EQ(1, (*niter).dimension());
+        EXPECT_EQ(size_t(1), (*niter).dimension());
         EXPECT_EQ(a.shape()[2], (*niter).shape()[0]);
         EXPECT_EQ(a(1, 1, 0), (*niter)(0));
         EXPECT_EQ(a(1, 1, 1), (*niter)(1));
@@ -83,7 +83,7 @@ namespace xt
         auto iter = axis_begin(a);
         ++iter;
 
-        EXPECT_EQ(2, (*iter).dimension());
+        EXPECT_EQ(size_t(2), (*iter).dimension());
         EXPECT_EQ(a.shape()[1], (*iter).shape()[0]);
         EXPECT_EQ(a.shape()[2], (*iter).shape()[1]);
